% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivdesc.R
\name{ivdesc}
\alias{ivdesc}
\title{Profiling compliers and non-compliers for instrumental variable analysis}
\usage{
ivdesc(
  X,
  D,
  Z,
  W = NULL,
  data = NULL,
  link = "probit",
  variance = FALSE,
  boot = FALSE,
  bootn = 1000,
  balance = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{X}{vector or unquoted variable name with numeric covariate}

\item{D}{vector or unquoted variable name with binary treatment}

\item{Z}{vector or unquoted variable name with binary instrument}

\item{W}{optional matrix or formula for first-stage regression}

\item{data}{optional data frame with variables}

\item{link}{'logit' or 'probit' link function for first-stage regression?}

\item{variance}{Calculate the variance of the covariate for each subgroup?}

\item{boot}{Replace all standard errors with bootstrap standard errors?}

\item{bootn}{number of bootstraps (ignored if \code{boot=FALSE} )}

\item{balance}{Run balance test?}

\item{verbose}{if TRUE, print first-stage regression summary}

\item{...}{additional arguments to be passed to \code{ivdesc_all}}
}
\value{
Returns a object \code{ivdesc} with estimates for each subgroup (\code{co}: complier, \code{nt}: never-taker, \code{at} : always-taker) and the full sample: 

\itemize{
	\item \code{mu} and \code{mu_se} : Mean of \code{X} and standard error
 \item \code{pi} and \code{pi_se}: Proportion of each subgroup in the sample and standard error 
 \item \code{var}: Variance of \code{X} (if \code{variance=TRUE})
}

Can be coerced to a proper \code{data.frame} using \code{as.data.frame}.
}
\description{
Estimates the mean and variance of a covariate for the complier, never-taker and always-taker subpopulation.
}
\details{
This function estimates the mean and the associated standard error of \code{X} for the complier, never-taker and always-taker subpopulation within a sample where some, but not all, units are encouraged by instrument \code{Z} to take the treatment \code{D}. 
Observations with missing values in either \code{X}, \code{D},  \code{Z}, or \code{W} are droppped (listwise deletion). 

One-sided noncompliance is supported. The mean for the always-/never-taker subpopulation will only be computed if there are at least two observed units in these subpopulations.

If additional covariates are defined via the parameter \code{W}, the package uses the kappa-estimator.  

The balance test is a t-test allowing for unequal variances.
}
\examples{
 

 # Example 1: Albertson/Lawrence (2009)
 # see Marbach/Hangartner (2019) for details/discussion

 library(icsw)
 data(FoxDebate)

 ivdesc(X=readnews,D=watchpro,Z=conditn, data=FoxDebate) 

 

 \donttest{

 # Example 2: JTPA Data

 library(haven)
 jtpa <- read_dta("http://fmwww.bc.edu/repec/bocode/j/jtpa.dta") 

 ivdesc(age, training, assignmt, data=jtpa, bootn=500)
 ivdesc(hispanic, training, assignmt, data=jtpa, boot=FALSE)

 }
 
 


}
\references{
Moritz Marbach and Dominik Hangartner. (2020). Profiling Compliers and Non-compliers for Instrumental Variable Analysis. \emph{Political Analysis}, 28(3), 435-444.
}
\seealso{
\code{\link[AER]{ivreg}}
}
